add TotalDesign page for Totalized Format components
