+++ src/components/Seguimiento/Seguimiento/CoilsDimensions.js
@@ -1,9 +1,11 @@
 import { useEffect, useState } from "react"
+import { useAuth } from "../../../context/auth-context";
 import { colors } from "../../../styles/colors";
 import { Wrapper, Table, Td, Td1, Td_datos } from "./styles"
 
-export function CoilsDimentions(){
- 
+export function CoilsDimensions(){
+	const {originalDimensions, setOriginalDimensions, filas, setFilas,} = useAuth();
+
 	return(
 		<div>
 			<p style={{color: colors.font.text, fontWeight: "bold", marginBottom:10 }}>ORIGINAL</p>
@@ -11,19 +13,19 @@ export function CoilsDimentions(){
 				<Table>
 					<tr>
 						<Td> Alambre </Td>
-						<Td_datos> 111 </Td_datos>
+						<Td_datos> {originalDimensions.d} </Td_datos>
 					</tr>
 					<tr>
 						<Td> Diam.exterior 1 </Td>
-						<Td_datos> 111 </Td_datos>
+						<Td_datos> {originalDimensions.Dext} </Td_datos>
 					</tr>
 					<tr>
 						<Td> Diam.exterior 2 </Td>
-						<Td_datos> 111 </Td_datos>
+						<Td_datos> {originalDimensions.Dext2} </Td_datos>
 					</tr>
 					<tr>
 						<Td> Sentido de vueltas </Td>
-						<Td_datos> 111 </Td_datos>
+						<Td_datos> 0 </Td_datos>
 					</tr>
 					<tr>
 						<Td> Rigidez teorica </Td>
@@ -38,19 +40,19 @@ export function CoilsDimentions(){
 				<Table>
 				    <tr>
 						<Td> Longitud </Td>
-						<Td_datos> 111 </Td_datos>
+						<Td_datos> {originalDimensions.L0} </Td_datos>
 					</tr>
 					<tr>
 						<Td> Diam.interior 1 </Td>
-						<Td_datos> 111 </Td_datos>
+						<Td_datos> {originalDimensions.Dint1} </Td_datos>
 					</tr>
 					<tr>
 						<Td> Diam.interior 2 </Td>
-						<Td_datos> 111 </Td_datos>
+						<Td_datos> {originalDimensions.Dint2} </Td_datos>
 					</tr>
 					<tr>
 						<Td> Nro de vueltas </Td>
-						<Td_datos> 111 </Td_datos>
+						<Td_datos> {originalDimensions.N} </Td_datos>
 					</tr>
 					<tr>
 						<Td> LDA </Td>
@@ -74,11 +76,11 @@ export function CoilsDimentions(){
 					</tr>	
 				    <tr>
 						<Td1> Luz 1 </Td1>
-						<Td_datos> 00 </Td_datos>
+						<Td_datos> {originalDimensions.Luz1} </Td_datos>
 					</tr>
 					<tr>
 						<Td1> Vta red/ampl 1 </Td1>
-						<Td_datos> -- </Td_datos>
+						<Td_datos> {originalDimensions.Vtas1} </Td_datos>
 					</tr>
 					<tr>
 						<Td1> Excentricidad 1</Td1>
@@ -98,11 +100,11 @@ export function CoilsDimentions(){
 					</tr>	
 				    <tr>
 						<Td1> Luz 2 </Td1>
-						<Td_datos> 00 </Td_datos>
+						<Td_datos> {originalDimensions.Luz2} </Td_datos>
 					</tr>
 					<tr>
 						<Td1> Vta red/ampl 2 </Td1>
-						<Td_datos> -- </Td_datos>
+						<Td_datos> {originalDimensions.Vtas2} </Td_datos>
 					</tr>
 					<tr>
 						<Td1> Excentricidad 2</Td1>
