+++ src/components/Seguimiento/Seguimiento/styles.js
