--- src/pages/Design2/index.js
@@ -1,2 +1,0 @@
-import Design2 from "./Design2";
-export default Design2
