+++ src/pages/OriginalSample/index.js
@@ -1,0 +1,2 @@
+import OriginalSample from "./OriginalSample";
+export default OriginalSample
