+++ src/pages/OriginalSample/styles.js
